Renames some files to a name that makes more sense
