+++ gpg_group_chat/__init__.py
@@ -1,2 +1,2 @@
-from gpg_group_chat.client.app import Client
-from gpg_group_chat.server.app import Server
+from gpg_group_chat.client.client import Client
+from gpg_group_chat.server.server import Server
